requirements.txt
alembic==1.13.2
Babel==2.13.0
blinker==1.6.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email-validator==1.1.2
Flask==3.0.3
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Validator==1.4.2
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
isbnlib==3.10.3
iso3166==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.5
MarkupSafe==2.1.3
pillow==10.4.0
py-moneyed==0.8.0
pycountry==24.6.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
schwifty==2020.9.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2
